requirements.txt
paho-mqtt==1.6.1
schedule==1.1.0
